checkpoint - refactor engine
cleaning up for readability
